fix: Rename disk space config
## Context

Make disk space config more generic. 

## Objective

This PR renames the disk space config from `high` to `space`.

## References

Related to https://github.com/screwdriver-cd/executor-k8s-vm/pull/60

## License

<!-- The following line must be included in your pull request -->

I confirm that this contribution is made under the terms of the license found in the root directory of this repository's source tree and that I have the authority necessary to make this contribution on behalf of its copyright owner.